Implement `#hash` and `#eql?` in same terms
In our case we want to ensure two distinct value objects are considered
the same by hashes and sets when they have the same value.

When a object is added for the first time only the value `#hash` is
required to insert it into the hash table. However if the same object is
added, or used as a key, for a second time the hash will match and a
secondary equality check, using `#eql` will take place.

Therefore the desired behavior demands that both methods are implemented
in the same terms. We have a chosen a combination of the value, the
domain object, and a constant, the value object class, to make a
globally consistent hash value.

+@mikejustdoit