Handle many to one with a `nil` foreign key
Bypass proxy build and return nil then translate the nil association into [nil]
so that it is explicitly goes through the dump / serialization process.

This allows the nil foreign key to be extracted and merged into the
record.

Includes beginning of a graph serializer refactor.

`nil` has proven to be a huge pain in the assm, no surprises there.
Considering removing nils altogether and using collections of one /
an option type.